requirements.txt
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
Levenshtein==0.21.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
platformdirs==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.1
pytz==2023.3.post1
rapidfuzz==3.2.0
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
threadpoolctl==3.2.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
